Move solution to concurrency lib
